pyproject.toml
[project]
name = "contextguard"
version = "0.1.0"
description = "IaC Attack Path Prioritizer — contextual risk scoring for Terraform plans"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.scripts]
contextguard = "contextguard.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
disallow_any_generics = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
